move revision_view test to its own package as it caused flakes due to environment setup
